Upstream fixes (#102)
* More work

* temp fix

* Comment out Chroma push in copy.ps1

Don't attempt to push Chroma when pushing Noodle to the headset.
I am trying to get Noodle basics working first, I'll deal with Chroma later. For now, I don't have a Chroma 1.40.8 build so there's nothing to push.

* Bump for 1.40.8

Crashes upon starting a song, but at least the mod loads.

* Update / add ndk-stack, pull-tombstone, and addr2line scripts

Updated and added the scripts from Laurie's template that I use for debugging mods.

* Update build system to use a more recent version of Laurie's template

Honestly wasn't going to do this, but then I ran into an issue where none of my changes were affecting anything, and I jumped to the conclusion that it was the build process that was acting awry. Turns out the copy script puts the .so in a different folder and was acting as a duplicate to my manually placed .so file. Oh well.

* Revert "Update build system to use a more recent version of Laurie's template"

This reverts commit 1d80f7a6fec5996e0f8eba24a44115fb3399b110.
For some reason it builds as v0.0.0 and the play button is once again disabled... I don't really feel like figuring this one out right now.

* Update scripts

Updating the cmake build files caused issues, but it seems the scripts still work. I use copy.ps1, as it functions in Laurie's template, a lot.

* Update README, add 1.40.8 to-do list

* Update README, add new findings to to-do list

* Add logging to track down jump animation issue

* Fix note jump animations

NoodleMovementDataProvider::get_moveDuration should return the original value rather than 0. Other numbers get divided by this afterall.
This was a parity issue with PC.

* Revert "Add logging to track down jump animation issue"

This reverts commit b2b31a4f21cfd04eb4cb9b250746f1961c5e4a0e.

* Update README, adding new bugs

Now that tracks appear, all the other bugs have become visible! I'm limiting myself to Chroma-less maps right now, greatly reducing the pool of maps to find bugs in.

* Sync NoodleMovementDataProvider to PC version

* Fix note rotation

* Copy missed time logic from PC

* Fix a typo and use move duration from the game

* Fix note jumps

* Fix crash on loading second noodle map

* Fix crashes on song restarts and fix fake obstacle head collision handling

* Update scripts

* Remove log.txt and ignore mod.json

* Update actions

* Update cmakelists and qpm scripts

---------

Co-authored-by: Futuremapper <54294576+Futuremappermydud@users.noreply.github.com>
Co-authored-by: CatsaCode <catsacode@gmail.com>